requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
bitmex==0.2.2
bravado==10.3.2
bravado-core==5.12.0
cbpro==1.1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
dateparser==0.7.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jsonpointer==2.0
jsonref==0.2
jsonschema==3.0.1
monotonic==1.5
msgpack-python==0.5.6
numpy==1.16.3
pandas==0.24.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyHamcrest==1.9.0
pymongo==3.5.1
pyOpenSSL==19.0.0
pyrsistent==0.14.11
python-binance==0.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
regex==2019.4.14
requests==2.13.0
rfc3987==1.3.8
service-identity==18.1.0
simplejson==3.16.0
six==1.10.0
sortedcontainers==2.1.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
Twisted==19.2.0
txaio==18.8.1
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.24.2
webcolors==1.8.1
websocket-client==0.40.0
zope.interface==4.6.0
